NODE-412: Code review: Fix compilation errors after moving stuff to 'synchronization' package
