+++ comm/src/main/scala/io/casperlabs/comm/gossiping/synchronization/InitialSynchronization.scala
@@ -1,4 +1,4 @@
-package io.casperlabs.comm.gossiping
+package io.casperlabs.comm.gossiping.synchronization
 
 import cats.Parallel
 import cats.effect._
@@ -9,9 +9,9 @@ import io.casperlabs.casper.consensus.BlockSummary
 import io.casperlabs.catscontrib.MonadThrowable
 import io.casperlabs.comm.discovery.NodeUtils.showNode
 import io.casperlabs.comm.discovery.{Node, NodeDiscovery}
-import io.casperlabs.comm.gossiping.InitialSynchronization.SynchronizationError
 import io.casperlabs.comm.gossiping.Utils.hex
-import io.casperlabs.comm.gossiping.synchronization.Synchronizer
+import io.casperlabs.comm.gossiping.synchronization.InitialSynchronization.SynchronizationError
+import io.casperlabs.comm.gossiping._
 import io.casperlabs.models.BlockImplicits._
 import io.casperlabs.shared.Log
 import io.casperlabs.shared.IterantOps.RichIterant
